requirements.txt
botbuilder-core>=4.14
botbuilder-integration-aiohttp>=4.14
langgraph>=1.1
langgraph-checkpoint-sqlite>=3.0
langchain-google-genai>=4.0
langchain-core>=1.2
google-cloud-logging>=3.0
google-cloud-monitoring>=2.0
google-cloud-run>=0.10
google-cloud-resource-manager>=1.0
PyGithub>=2.0
PyJWT[crypto]>=2.0
httpx>=0.28
aiohttp>=3.9
aiosqlite>=0.20
numpy>=2.0
